requirements.txt
openai>=1.0.0
pyyaml>=6.0
aiosqlite>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
jinja2>=3.1.0
